Migrated isGraphiQLEnabled and isVoyagerEnabled
