+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/EndpointAnnotators/ClientEndpointAnnotatorInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GraphQLAPI\GraphQLAPI\Services\EndpointAnnotators;
+
+use WP_Post;
+
+interface ClientEndpointAnnotatorInterface extends EndpointAnnotatorInterface
+{
+    public function isClientEnabled(WP_Post|int $postOrID): bool;
+}
